#167 ContentHelper moved to AemCommon module, DataPathUtil adjusted
