+++ bb-aem-touch-ui/src/main/java/com/cognifide/qa/bb/aem/touch/pageobjects/Component.java
@@ -19,10 +19,8 @@
  */
 package com.cognifide.qa.bb.aem.touch.pageobjects;
 
-import static com.cognifide.qa.bb.aem.touch.util.ContentHelper.JCR_CONTENT;
 import static org.openqa.selenium.support.ui.ExpectedConditions.visibilityOf;
 
-import org.apache.commons.lang3.StringUtils;
 import org.openqa.selenium.NoSuchElementException;
 import org.openqa.selenium.StaleElementReferenceException;
 import org.openqa.selenium.WebElement;
@@ -30,6 +28,7 @@ import org.openqa.selenium.WebElement;
 import com.cognifide.qa.bb.aem.touch.data.componentconfigs.ComponentConfiguration;
 import com.cognifide.qa.bb.aem.touch.pageobjects.dialogs.ConfigDialog;
 import com.cognifide.qa.bb.aem.touch.pageobjects.dialogs.DeleteDialog;
+import com.cognifide.qa.bb.aem.touch.util.DataPathUtil;
 import com.cognifide.qa.bb.constants.HtmlTags;
 import com.cognifide.qa.bb.provider.selenium.BobcatWait;
 import com.cognifide.qa.bb.qualifier.CurrentScope;
@@ -72,7 +71,7 @@ public class Component {
   public String getDataPath() {
     String rawValue =
         bobcatWait.staleSafe(currentScope, checked -> checked.getAttribute(HtmlTags.Attributes.DATA_PATH));
-    return StringUtils.substringAfter(rawValue, JCR_CONTENT);
+    return DataPathUtil.extract(rawValue);
   }
 
   /**
